chore: move certs to a single file; use constant for bit size
