--- internal/pkg/certs/ca.go
@@ -1,53 +1,0 @@
-package certs
-
-import (
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"fmt"
-	"math/big"
-	"time"
-)
-
-func GenerateCA() ([]byte, *rsa.PrivateKey, error) {
-	serialNumber, err := rand.Int(rand.Reader, (&big.Int{}).Exp(big.NewInt(2), big.NewInt(159), nil))
-	if err != nil {
-		return nil, nil, fmt.Errorf("cannot init serial number: %w", err)
-	}
-
-	privateKey, err := rsa.GenerateKey(rand.Reader, 4096)
-	if err != nil {
-		return nil, nil, fmt.Errorf("cannot create private key: %w", err)
-	}
-
-	caCert := x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			Organization:  []string{""},
-			Country:       []string{""},
-			Province:      []string{""},
-			Locality:      []string{""},
-			StreetAddress: []string{""},
-			PostalCode:    []string{""},
-		},
-		NotBefore:             time.Now(),
-		NotAfter:              time.Now().AddDate(10, 0, 0),
-		IsCA:                  true,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageCertSign,
-		BasicConstraintsValid: true,
-	}
-
-	caCertBytes, err := x509.CreateCertificate(
-		rand.Reader,
-		&caCert,
-		&caCert,
-		&privateKey.PublicKey,
-		privateKey)
-	if err != nil {
-		return nil, nil, fmt.Errorf("cannot create certificate: %w", err)
-	}
-
-	return caCertBytes, privateKey, nil
-}
